+++ packages/graphql/src/Builder/Sources/InterfaceFieldArgumentSource.php
@@ -13,11 +13,14 @@ use GraphQL\Type\Definition\FieldDefinition;
 use GraphQL\Type\Definition\InterfaceType;
 use GraphQL\Type\Definition\Type;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Manipulator;
+use LastDragon_ru\LaraASP\GraphQL\Builder\Sources\Traits\FieldArgument;
 
 /**
  * @extends Source<NamedTypeNode|ListTypeNode|NonNullTypeNode|Type>
  */
 class InterfaceFieldArgumentSource extends Source {
+    use FieldArgument;
+
     public function __construct(
         Manipulator $manipulator,
         private InterfaceTypeDefinitionNode|InterfaceType $object,
@@ -39,18 +42,6 @@ class InterfaceFieldArgumentSource extends Source {
 
     public function getArgument(): InputValueDefinitionNode|Argument {
         return $this->argument;
-    }
-    // </editor-fold>
-
-    // <editor-fold desc="TypeSource">
-    // =========================================================================
-    public function __toString(): string {
-        $manipulator = $this->getManipulator();
-        $argument    = $manipulator->getNodeName($this->getArgument());
-        $field       = $manipulator->getNodeName($this->getField());
-        $type        = $manipulator->getNodeTypeFullName($this->getObject());
-
-        return "{$type} { {$field}({$argument}) }";
     }
     // </editor-fold>
 }
