+++ packages/graphql/src/Builder/Sources/Traits/Field.php
@@ -1,0 +1,21 @@
+<?php declare(strict_types = 1);
+
+namespace LastDragon_ru\LaraASP\GraphQL\Builder\Sources\Traits;
+
+use LastDragon_ru\LaraASP\GraphQL\Builder\Sources\InputFieldSource;
+use LastDragon_ru\LaraASP\GraphQL\Builder\Sources\InterfaceFieldSource;
+use LastDragon_ru\LaraASP\GraphQL\Builder\Sources\ObjectFieldSource;
+
+/**
+ * @mixin InputFieldSource|InterfaceFieldSource|ObjectFieldSource
+ * @internal
+ */
+trait Field {
+    public function __toString(): string {
+        $manipulator = $this->getManipulator();
+        $field       = $manipulator->getNodeName($this->getField());
+        $type        = $manipulator->getNodeTypeFullName($this->getObject());
+
+        return "{$type} { {$field} }";
+    }
+}
